requirements.txt
duckdb==1.4.4
streamlit==1.54.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
requests==2.32.5
PyYAML==6.0.3
Pillow==10.4.0
scipy==1.15.3
python-dateutil==2.9.0.post0
tenacity==8.5.0
